requirements.txt
flask==2.3.3
gunicorn==21.2.0
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-image==0.21.0
numpy==1.24.3
Pillow==10.0.1
easyocr==1.7.0
